changed Mapping and included all the reports logic

+++ src/main/java/com/java/finance/ThirukumaranFinance/Entity/DailyTotal.java
@@ -25,7 +25,7 @@ public class DailyTotal {
 	private int id;
 	
 	@Column(name = "total_amount")
-	private long totalAmount;
+	private int totalAmount;
 	
 	@Column(name = "date")
 	private LocalDate date;
